% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_microbiota.R
\name{plot_gut_microbiota}
\alias{plot_gut_microbiota}
\title{Clustered stackbar coupled with DESeq2}
\usage{
plot_gut_microbiota(
  ps_object = ps_unfiltered,
  exp_group = "group",
  subset_group = NULL,
  sample_name = "SampleID",
  main_level = "Phylum",
  sub_level = "Family",
  threshold = 1,
  n_phy = 4,
  hues = c("Oranges", "Greens", "Blues", "Purples"),
  color_bias = 2,
  text_size = 9,
  legend_size = 7,
  x_axis_size = 8,
  differential_analysis = T,
  test = c("Wald", "LRT")[1],
  fitType = c("parametric", "local", "mean", "glmGamPoi")[1],
  sfType = c("ratio", "poscounts", "iterate")[1],
  betaPrior = FALSE,
  reduced = FALSE,
  quiet = FALSE,
  minReplicatesForReplace = 7,
  modelMatrixType = c("standard", "expanded")[1],
  useT = FALSE,
  minmu = if (fitType == "glmGamPoi") 1e-06 else 0.5,
  parallel = FALSE,
  BPPARAM = bpparam(),
  fdr_threshold = 0.05,
  sig_lab = T,
  ...
)
}
\arguments{
\item{ps_object}{A \linkS4class{phyloseq} object. Must have a \link{sample_data}, \link{otu_table}, and \link{tax_table} components.}

\item{exp_group}{A column in the \link{sample_data} containing the experimental group information.}

\item{subset_group}{Default NULL. Groups among the 'exp_group' column from the \link{sample_data} object to subset.}

\item{sample_name}{Name of the column in \link{sample_data} containing the unique sample identifier.}

\item{main_level}{Default 'Phylum'. Level present in the \link{tax_table} to which taxa will be clustered.}

\item{sub_level}{Default 'Family'. Level present in the \link{tax_table} to which taxa will be plotted and analyzed.}

\item{threshold}{Default 1\%, threshold to regroup taxa with lower relative abundance into the 'other' groups}

\item{n_phy}{Default 4, number of main_level to plot. Same number of colors must be given in the 'hues' parameter.}

\item{hues}{Color used to represent main_level, should be the same number than n_phy parameter. See \link{colorRampPalette}.}

\item{color_bias}{Define the gradiant of shades among the colors. See \link{colorRampPalette}.}

\item{text_size}{Control the text size of the graph. See \link{ggplot2}.}

\item{legend_size}{Control the legend text size of the graph. see ggplot2. See \link{ggplot2}.}

\item{x_axis_size}{Control the x axis text size of the graph. see ggplot2. See \link{ggplot2}.}

\item{differential_analysis}{Default TRUE. Whether or not use DESeq2 to do a differential abundance analysis. See \link{DESeq2-package}.}

\item{test}{Default 'Wald'. See \link{DESeq2-package}.}

\item{fitType}{Default "parametric". See \link{DESeq2-package}.}

\item{sfType}{Default "ratio". See \link{DESeq2-package}.}

\item{betaPrior}{Default FALSE.See \link{DESeq2-package}.}

\item{reduced}{Default FALSE. See \link{DESeq2-package}.}

\item{quiet}{Default FALSE. See \link{DESeq2-package}.}

\item{minReplicatesForReplace}{Default '7'. See \link{DESeq2-package}.}

\item{modelMatrixType}{Default "standard". See \link{DESeq2-package}.}

\item{useT}{Default FALSE. See \link{DESeq2-package}.}

\item{minmu}{See \link{DESeq2-package}.}

\item{parallel}{Default FALSE. See \link{DESeq2-package}.}

\item{BPPARAM}{See \link{DESeq2-package}.#'}

\item{sig_lab}{Default TRUE. Wheter add stars after taxa name reflecting statistical significance.}

\item{...}{additionnal parameters passed into the \code{\link{phyloseq_to_deseq2}} functions, see \link{Phyloseq}.}
}
\value{
A \link{list} containing  \link{ggplot2} '$plot' object and, if differential_analysis = TRUE, a \link{dataframe} from \code{\link{results}} with DESeq2 outputs of significant features '$significant_table'
}
\description{
Create a plot of taxa relative abundance, clustered by phylogeny with the possibility to highlight the differentially abundant features.
}
\examples{

\dontrun{
data(GlobalPatterns)
ps_unfiltered <- GlobalPatterns

my_plot <- plot_gut_microbiota(ps_object = ps_unfiltered,
exp_group = 'SampleType',
subset_group = c("Feces", "Skin"),
sample_name = 'X.SampleID',
differential_analysis = T,
sig_lab = T,
fdr_threshold = 0.8)

print(my_plot$plot)
print(my_plot$significant_table)

}

}
\seealso{
\link{phyloseq}

\link{DESeq2-package}

\link{ggplot2}
}
\author{
Thibault Cuisiniere, Marco Constante, Manuela M. Santos
}
