% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustered_stackbar.R
\name{plot_gut_microbiota}
\alias{plot_gut_microbiota}
\title{Clustered stackbar coupled with DESeq2}
\usage{
plot_gut_microbiota(
  ps_object = ps_unfiltered,
  exp_group = "group",
  subset_group = NULL,
  sample_name = "SampleID",
  main_level = "Phylum",
  sub_level = "Family",
  threshold = 1,
  n_phy = 4,
  hues = c("Oranges", "Greens", "Blues", "Purples"),
  color_bias = 2,
  text_size = 9,
  legend_size = 7,
  x_axis_size = 8,
  animate = F,
  differential_analysis = T,
  test = c("Wald", "LRT")[1],
  fitType = c("parametric", "local", "mean", "glmGamPoi")[1],
  sfType = c("ratio", "poscounts", "iterate")[1],
  betaPrior = FALSE,
  reduced = FALSE,
  quiet = FALSE,
  minReplicatesForReplace = 7,
  modelMatrixType = c("standard", "expanded")[1],
  useT = FALSE,
  minmu = if (fitType == "glmGamPoi") 1e-06 else 0.5,
  parallel = FALSE,
  BPPARAM = bpparam(),
  fdr_threshold = 0.05,
  ...
)
}
\arguments{
\item{ps_object}{a phyloseq object. \code{\link{phyloseq-class}}.
Must have a \code{\link{sample_data}} component.}

\item{exp_group}{a column in sample_data containing the experimental group information}

\item{subset_group}{default NULL. Groups among the 'exp_group' column to subset}

\item{sample_name}{name of the column in sample_data containing the unique sample identifier}

\item{main_level}{Default 'Phylum'. Level to which features will be clustered}

\item{sub_level}{Default 'Family'. Level to which taxa will be plotted and analyzed}

\item{threshold}{Default 1\%, threshold to regroup taxa with lower relative abundance into the 'other' groups}

\item{n_phy}{Default 4, number of main_level to plot. Same number of color must be given in the 'hues' parameter.}

\item{hues}{Color used to represent main_level, should be the same number than n_phy parameter.}

\item{color_bias}{Define the gradiant of shades among the colors}

\item{text_size}{control the text size. see ggplot2}

\item{legend_size}{control the legend text size. see ggplot2}

\item{x_axis_size}{control the x axis text size. see ggplot2}

\item{differential_analysis}{Default TRUE. Whether or not use DESeq2 to do a differential abundance analysis}

\item{test}{Default 'Wald'. see DESeq2}

\item{fitType}{Default "parametric", see DESeq2}

\item{sfType}{Default "ratio",see DESeq2}

\item{betaPrior}{Default FALSE,see DESeq2}

\item{reduced}{Default FALSE,see DESeq2}

\item{quiet}{Default FALSE,see DESeq2}

\item{minReplicatesForReplace}{Default '7',see DESeq2}

\item{modelMatrixType}{Default "standard",see DESeq2}

\item{useT}{Default FALSE,see DESeq2}

\item{minmu}{see DESeq2}

\item{parallel}{Default FALSE,see DESeq2}

\item{BPPARAM}{see DESeq2}

\item{...}{additionnal parameters passed into the phyloseq_to_deseq2 functions, see Phyloseq}
}
\value{
A list containing  ggplot $plot, object and if differential_analysis = TRUE, a dataframe with DESeq2 outputs of significant features $significant_table
}
\description{
Create a plot of taxa relative abundance, clustered by phylogeny with the possibility to highlight the differentially abundant features.
}
\examples{
data(GlobalPatterns)
ps_unfiltered <- GlobalPatterns
my_plot <- plot_gut_microbiota(ps_object = ps_unfiltered,
exp_group = 'SampleType',
subset_group = c("Feces", "Skin"),
sample_name = 'X.SampleID',
differential_analysis = T,
fdr_threshold = 0.8)
print(my_plot$plot)
print(my_plot$significant_table)

}
\author{
Thibault Cuisiniere, Marco Constante, Manuela M. Santos
}

